collecting_data/ver_one/app/Scripts/WHENET_TRT/requirements.txt
onnx
pycuda>=2019.1.1
nvidia-pyindex
nvidia-tensorrt